+++ SalesStatistics/SalesStatistics/Migrations/SalesDbContextModelSnapshot.cs
@@ -31,8 +31,8 @@ namespace SalesStatistics.Migrations
                         .HasMaxLength(40)
                         .HasColumnType("nvarchar(40)");
 
-                    b.Property<double>("Cost")
-                        .HasColumnType("float");
+                    b.Property<int>("Cost")
+                        .HasColumnType("int");
 
                     b.Property<DateTime>("Date")
                         .HasColumnType("datetime2");
